Created a new commands subdirectory for python command files that add subparsers to main args parser
