+++ project_generator/main.py
@@ -15,15 +15,10 @@
 import argparse
 import os
 import logging
+
 import pkg_resources
 
-from . import init
-from . import clean
-from . import export
-from . import update
-from . import list_projects
-from . import build
-from . import flash
+from project_generator.commands import build, clean, export, flash, init, list_projects, update
 
 subcommands = {
     'init': init,
